feature:班级容量和导入名单时初始化

+++ src/main/java/cn/homyit/onlineLeaveSystem/eneity/DO/SysStudentClassInfo.java
@@ -18,7 +18,7 @@ public class SysStudentClassInfo implements Serializable {
   private Long id;
   private String majorAndClass;
 
-  private Long capacity;
+  private Integer capacity;
   private String highUser;
   @TableLogic
   private Long deleted;
